est_cpu_minus_tracker_time change
Added a check to make sure that the tracker is actually connected before
attempting to get times from it.

Also added in the codes/states/descriptions as a docstring in the
trackerstate property.

Also created a new error code to be used.

Also added in *1.0 to the numbers at the end to make sure we are doing
float division.